fix: Fixed db migrations

+++ migrations/sqls/20210826050734-orders-up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE orders (
     id SERIAL PRIMARY KEY,
-    user_id bigint REFERENCES users(id),
+    user_id integer REFERENCES users(id),
     status VARCHAR(64)
 );
